Add pretty-print column and formatters
The recorder now also adds a table column (`pretty_print`) for pretty-printed
transformation of the received message. If a pretty-print formatter is
specified, that column is populated with a formatted version of the
payload. This makes it easier to use the database for manual debugging
or monitoring of the Zyre nodes.

Also added a section to README about the formatters.